% --- Source file: ExactHW-package.Rd ---
\name{GWASExactHW-package}
\alias{GWASExactHW-package}
\alias{GWASExactHW}
\docType{package}
\title{Exact Hardy-Weinburg testing for Genome Wide Association Studies}
\description{
This package contains a function to do exact Hardy-Weinburg testing (using Fisher's test) over all or a selection of SNP genotypes as typically obtained in a Genome Wide Association Study (GWAS).}
\details{
\tabular{ll}{
Package: \tab GWASExactHW\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-08-24\cr
License: \tab GNU \cr
LazyLoad: \tab yes\cr
}
The function HWExact runs fast Hardy-Weinburg testing for a set of bi-alleleic genotypes.
}
\author{Ian Painter, GENEVA project, University of Washington

Maintainer: Ian Painter <ipainter@u.washington.edu>

}
\references{ Wigginton, JE, Cutler, DJ, and Abecasis, GR (2005) A Note on Exact Tests of Hardy-Weinberg Equilibrium. American Journal of Human Genetics. 76
}

\keyword{ package }
\seealso{

}
\examples{
}